+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.1
+- Fix the manual element resize edit boxes not properly saving their values
+- Make background layers independently moveable 
+
 ## 0.2.0
 - Split `detail_overlay` layer into 3 sprites: `health_foreground`, `stamina_foreground` `mana_foreground`